Src/requirements.txt
certifi==2021.10.8
chardet==4.0.0
cx-Freeze==6.8.1
cx-Logging==3.0
idna==2.10
importlib-metadata==4.8.1
Pillow==8.4.0
Pmw==2.0.1
requests==2.25.1
tkinterdnd2==0.3.0
urllib3==1.26.7
witkets==2.0
zipp==3.6.0
